fix: column improvements [ci]
* Remove name from typed column. This was not correct, a column is implicitly named if none is provided. If you need to ensure a column is named, then create it with one, or alias it, its not a property of a column.
* Implicit conversion from dotnet type to typed column. This allow for the removal of all doubled up functions that operated on a column and a native type. Now native types just auto cast to the respective column type. I have also double cast the native types that are not currently supported via the JVM bridge to ones that are. This is not 100% safe, but is better than no support at all.

The only API change that is not 100% great, to ensure the base function implementations on column can always produce new instances of their this type, I have added in a new() constraint. A future enhancement might be (if we keep this design) add an analyser rule to stop users calling new on typed columns. Instead they should always use the New static constructor functions.